Implement first pass at Interactivity live code example
The live code example below is a header navigation menu that enables interactivity using the
`:hover` CSS pseudo selector. On hover of a menu item, the text goes from dark to light, the
item background image goes from light to dark, and the icon goes from a "dormant" to "active"
state.

Also:
- Add index page
- Add Font Awesome icon font